#8: migrate samples to X impl, fix bugs

+++ metrics-facade-x/src/main/java/com/ringcentral/platform/metrics/x/histogram/hdr/HdrXHistogramImplSnapshot.java
@@ -79,7 +79,7 @@ public class HdrXHistogramImplSnapshot implements XHistogramImplSnapshot {
             return NO_VALUE;
         }
 
-        double upperBound = bucket.upperBoundAsLong();
+        long upperBound = bucket.upperBoundAsLong();
 
         for (int i = 0; i < bucketUpperBounds.length; ++i) {
             if (upperBound <= bucketUpperBounds[i]) {
